# Enable profile guided optimization for beacon-chain