--- a/beacon-chain/db/kv/init_test.go
+++ b/beacon-chain/db/kv/init_test.go
@@ -1,7 +1,12 @@
 package kv
 
 import (
+	"io"
+	"os"
+	"testing"
+
 	"github.com/prysmaticlabs/prysm/v4/config/params"
+	"github.com/sirupsen/logrus"
 )
 
 func init() {
@@ -9,4 +14,10 @@ func init() {
 	if err := params.SetActive(params.MainnetTestConfig()); err != nil {
 		panic(err)
 	}
+}
+
+func TestMain(m *testing.M) {
+	logrus.SetLevel(logrus.DebugLevel)
+	logrus.SetOutput(io.Discard)
+	os.Exit(m.Run())
 }