… beacon-chain/BUILD.bazel
… beacon-chain/operations/attestations/kv/benchmark_test.go
… cmd/beacon-chain/pprof.beacon-chain.samples.cpu.pb.gz
… testing/benchmark/benchmark_files/fullBlock-128Atts-16384Vals.ssz
… validator/BUILD.bazel