**What type of PR is this?**

Feature

**What does this PR do? Why is it needed?**

Net improvement is  2.19% performance boost in CPU. This will likely be even better with future versions of go.

<details>
<summary>benchmark stats</summary>

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7950X 16-Core Processor            
                                                                                          │  /tmp/old.txt   │              /tmp/new2.txt              │
                                                                                          │     sec/op      │     sec/op       vs base                │
Hash-32                                                                                        2.448 ± 102%      2.450 ± 102%        ~ (p=0.441 n=20)
HashMP-32                                                                                     94.98m ±   1%     93.93m ±   1%   -1.10% (p=0.004 n=10)
MutexRead-32                                                                                  2.812n ±   3%     2.885n ±   3%        ~ (p=0.101 n=10)
AtomicValueRead-32                                                                           0.3916n ±   4%    0.3890n ±   5%        ~ (p=0.912 n=10)
AtomicBoolRead-32                                                                            0.2041n ±  10%    0.2026n ±   7%        ~ (p=0.928 n=10)
MutexWrite-32                                                                                 2.815n ±   3%     2.889n ±   3%        ~ (p=0.210 n=10)
AtomicValueWrite-32                                                                           3.074n ±   1%     2.888n ±   4%   -6.05% (p=0.000 n=10)
AtomicBoolWrite-32                                                                            1.449n ±   4%     1.492n ±   4%        ~ (p=0.305 n=10)
MutexCAS-32                                                                                   6.082n ±   1%     6.060n ±   2%        ~ (p=0.989 n=10)
AtomicBoolCAS-32                                                                              1.618n ±   3%     1.617n ±   3%        ~ (p=1.000 n=10)
MutexToggle-32                                                                                6.079n ±   2%     6.153n ±   2%        ~ (p=0.093 n=10)
AtomicBoolToggle-32                                                                           1.448n ±   3%     1.437n ±   3%   -0.76% (p=0.050 n=10)
FeedSend1000-32                                                                               283.1µ ±   5%     271.0µ ±   2%   -4.28% (p=0.004 n=10)
HasBlockDB-32                                                                                 227.4n ±  21%     219.5n ±  21%        ~ (p=0.684 n=10)
HasBlockForkChoiceStore_DoublyLinkedTree-32                                                   11.41n ±   2%     11.15n ±   1%   -2.24% (p=0.000 n=10)
DepositTree_InsertNewImplementation-32                                                        4.627m ±   2%     4.629m ±   1%        ~ (p=0.853 n=10)
DepositTree_InsertOldImplementation-32                                                        91.83m ±   2%     89.48m ±   3%   -2.56% (p=0.003 n=10)
DepositTree_HashTreeRootNewImplementation-32                                                  292.0µ ±   4%     276.1µ ±   3%   -5.44% (p=0.000 n=10)
DepositTree_HashTreeRootOldImplementation-32                                                  251.9n ±   3%     236.0n ±   3%   -6.31% (p=0.000 n=10)
BellatrixComplete-32                                                                          565.2n ±   3%     550.4n ±  10%        ~ (p=0.143 n=10)
ComputeCommittee300000_WithPreCache-32                                                        276.9m ±   1%     271.7m ±   2%   -1.86% (p=0.000 n=10)
ComputeCommittee3000000_WithPreCache-32                                                       276.4m ±   2%     269.8m ±   1%   -2.39% (p=0.000 n=10)
ComputeCommittee128000_WithOutPreCache-32                                                     276.0m ±   2%     275.0m ±   1%        ~ (p=0.190 n=10)
ComputeCommittee1000000_WithOutCache-32                                                       276.3m ±   4%     273.3m ±   1%   -1.08% (p=0.002 n=10)
ComputeCommittee4000000_WithOutCache-32                                                       278.1m ±   1%     272.3m ±   1%   -2.09% (p=0.005 n=10)
ShuffledIndex/ShuffledIndex_4000000-32                                                        93.03µ ±   2%     90.38µ ±   1%   -2.85% (p=0.005 n=10)
ShuffledIndex/ShuffledIndex_40000-32                                                          92.78µ ±   2%     90.48µ ±   1%   -2.49% (p=0.019 n=10)
ShuffledIndex/ShuffledIndex_400-32                                                            93.36µ ±   1%     90.39µ ±   3%   -3.18% (p=0.000 n=10)
IndexComparison/Indexwise_ShuffleList_400000-32                                                36.85 ±   0%      35.86 ±   1%   -2.69% (p=0.000 n=10)
IndexComparison/Indexwise_ShuffleList_40000-32                                                 3.695 ±   1%      3.584 ±   1%   -3.02% (p=0.000 n=10)
IndexComparison/Indexwise_ShuffleList_400-32                                                  37.41m ±   2%     36.05m ±   1%   -3.63% (p=0.000 n=10)
ShuffleList/ShuffleList_400000-32                                                             421.6m ±   2%     263.3m ±   1%  -37.56% (p=0.000 n=10)
ShuffleList/ShuffleList_40000-32                                                              42.88m ±   3%     26.41m ±   4%  -38.40% (p=0.000 n=10)
ShuffleList/ShuffleList_400-32                                                                529.1µ ±   2%     360.2µ ±   4%  -31.92% (p=0.000 n=10)
ExecuteStateTransition_FullBlock-32                                                           58.57m ±   1%     57.80m ±   2%        ~ (p=0.089 n=10)
ExecuteStateTransition_WithCache-32                                                           59.27m ±   1%     58.43m ±   2%   -1.42% (p=0.001 n=10)
ProcessEpoch_2FullEpochs-32                                                                   6.096m ±   2%     5.897m ±   2%   -3.26% (p=0.015 n=10)
HashTreeRoot_FullState-32                                                                     293.5n ±   3%     276.9n ±   3%   -5.65% (p=0.001 n=10)
HashTreeRootState_FullState-32                                                                287.8n ±   7%     264.8n ±   5%   -7.96% (p=0.002 n=10)
MarshalState_FullState/Proto_Marshal-32                                                       3.756m ±   1%     3.842m ±   2%   +2.29% (p=0.000 n=10)
MarshalState_FullState/Fast_SSZ_Marshal-32                                                    1.226m ±   3%     1.235m ±   2%        ~ (p=0.315 n=10)
UnmarshalState_FullState/Proto_Unmarshal-32                                                   11.79m ±   3%     11.81m ±   2%        ~ (p=0.971 n=10)
UnmarshalState_FullState/Fast_SSZ_Unmarshal-32                                                6.529m ±   4%     6.273m ±   2%   -3.93% (p=0.004 n=10)
Store_BlobSidecarsByRoot-32                                                                   1.203m ±   2%     1.230m ±   1%   +2.21% (p=0.043 n=10)
State_CheckStateSaveTime_1-32                                                                 38.26m ±   2%     37.62m ±   2%        ~ (p=0.089 n=10)
State_CheckStateSaveTime_10-32                                                                50.79m ±  15%     49.14m ±  11%   -3.24% (p=0.035 n=10)
State_CheckStateReadTime_1-32                                                                 23.55m ±   4%     21.82m ±   7%   -7.33% (p=0.001 n=10)
State_CheckStateReadTime_10-32                                                                20.64m ±   9%     19.91m ±   2%   -3.57% (p=0.002 n=10)
Store_SaveLastValidatedCheckpoint-32                                                          2.516m ±  13%     2.510m ±   2%        ~ (p=0.739 n=10)
HighestAttestations-32                                                                         3.052 ±   1%      3.072 ±   3%        ~ (p=0.853 n=10)
Store_CheckDoubleBlockProposals-32                                                            355.8m ±   4%     350.2m ±   4%        ~ (p=0.353 n=10)
ExtractGossipDigest-32                                                                       11.010n ±   1%     9.442n ±   1%  -14.24% (p=0.000 n=10)
_retrieveIndicesFromBitfield-32                                                               244.1n ±   1%     243.5n ±   1%        ~ (p=0.446 n=10)
GetPeers-32                                                                                   4.902m ±   4%     4.898m ±   2%        ~ (p=0.631 n=10)
CommitteeAssignment-32                                                                        11.06m ±   4%     10.75m ±   4%   -2.79% (p=0.009 n=10)
Server_PrepareBeaconProposer-32                                                               6.469m ±  13%     6.158m ±  10%        ~ (p=0.123 n=10)
ProposerAtts_sortByProfitability/max-cover_256_attestations_with_single_bit_set-32            2.572m ±   4%     2.626m ±   4%        ~ (p=0.353 n=10)
ProposerAtts_sortByProfitability/max-cover_256_attestations_with_64_random_bits_set-32        2.537m ±   4%     2.624m ±   5%        ~ (p=0.247 n=10)
ProposerAtts_sortByProfitability/max-cover_512_attestations_with_single_bit_set-32            8.361m ±   3%     8.272m ±   5%        ~ (p=0.280 n=10)
ProposerAtts_sortByProfitability/max-cover_1024_attestations_with_64_random_bits_set-32       6.246m ±   2%     6.158m ±   4%        ~ (p=0.190 n=10)
ProposerAtts_sortByProfitability/max-cover_1024_attestations_with_512_random_bits_set-32      1.942m ±   2%     1.931m ±   2%        ~ (p=0.218 n=10)
ProposerAtts_sortByProfitability/max-cover_1024_attestations_with_1000_random_bits_set-32     1.475m ±   2%     1.444m ±   2%   -2.15% (p=0.015 n=10)
CheckSlashableAttestations/1_attestation_1_validator-32                                       2.694m ±  43%     2.678m ±   3%        ~ (p=1.000 n=10)
CheckSlashableAttestations/1_attestation_100_validators-32                                    2.907m ±  16%     2.852m ± 279%        ~ (p=0.353 n=10)
CheckSlashableAttestations/1_attestation_1000_validators-32                                   13.66m ±  77%     13.27m ±   7%        ~ (p=0.063 n=10)
CheckSlashableAttestations/100_attestations_1_validator-32                                    4.235m ±   1%     4.186m ±   3%        ~ (p=0.143 n=10)
CheckSlashableAttestations/100_attestations_100_validators-32                                 25.05m ±   3%     24.07m ±   8%   -3.92% (p=0.029 n=10)
CheckSlashableAttestations/100_attestations_1000_validators-32                                220.7m ±   5%     218.8m ±   2%        ~ (p=0.393 n=10)
CheckSlashableAttestations/1000_attestations_1_validator-32                                   17.36m ±   0%     17.08m ±   3%   -1.59% (p=0.023 n=10)
CheckSlashableAttestations/1000_attestations_100_validators-32                                200.6m ±   4%     190.9m ±   6%   -4.83% (p=0.011 n=10)
CheckSlashableAttestations/1000_attestations_1000_validators-32                                1.852 ±   2%      1.802 ±   4%   -2.71% (p=0.029 n=10)
AppendPreviousEpochAttestations-32                                                            12.92µ ±   3%     12.59µ ±   4%        ~ (p=0.123 n=10)
AppendEth1DataVotes-32                                                                        9.948µ ±   3%     9.884µ ±   2%        ~ (p=0.529 n=10)
AppendHistoricalRoots-32                                                                      55.01m ± 156%     70.93m ±  51%  +28.94% (p=0.015 n=10)
AppendHistoricalSummaries-32                                                                  24.43m ±   6%     30.40m ±  18%  +24.44% (p=0.000 n=10)
ParticipationBits-32                                                                          1.804m ±   4%     1.541m ±  15%  -14.57% (p=0.001 n=10)
AppendBalance-32                                                                              18.84m ±  40%     21.03m ±   9%        ~ (p=0.075 n=10)
AppendInactivityScore-32                                                                      18.59m ±  44%     20.25m ±  12%        ~ (p=0.123 n=10)
BeaconState/Vectorized_SHA256-32                                                            0.01053n ±  10%   0.01003n ±  16%        ~ (p=1.000 n=10)
BeaconState/Current_SHA256-32                                                               0.03662n ±  43%   0.03706n ±   3%        ~ (p=0.184 n=10)
CloneValidators_Proto-32                                                                      3.326m ±   3%     3.177m ±   2%   -4.48% (p=0.001 n=10)
CloneValidators_Manual-32                                                                     922.5µ ±   2%     870.2µ ±   2%   -5.67% (p=0.000 n=10)
StateClone_Proto-32                                                                           3.774m ±   1%     3.768m ±   1%        ~ (p=0.436 n=10)
StateClone_Manual-32                                                                          2.196m ±   1%     2.101m ±   2%   -4.33% (p=0.000 n=10)
Merkleize_Buffered-32                                                                         676.1µ ±   2%     676.3µ ±   1%        ~ (p=0.495 n=20)
Reference_MinusRef-32                                                                         6.146n ±   2%     6.125n ±   2%        ~ (p=0.280 n=10)
HashTreeRoot_Generic_512-32                                                                   7.411m ±   2%     7.338m ±   1%        ~ (p=0.247 n=10)
HashTreeRoot_Generic_16384-32                                                                 19.96m ±   1%     20.03m ±   2%        ~ (p=0.631 n=10)
HashTreeRoot_Generic_300000-32                                                                251.2m ±   1%     252.8m ±   1%        ~ (p=0.063 n=10)
ReturnTrieLayer_NormalAlgorithm-32                                                            1.050m ±   1%     1.045m ±   1%        ~ (p=0.063 n=10)
ReturnTrieLayer_VectorizedAlgorithm-32                                                        1.049m ±   1%     1.042m ±   0%   -0.65% (p=0.029 n=10)
ReturnTrieLayerVariable_NormalAlgorithm-32                                                    838.6µ ±   1%     814.0µ ±   0%   -2.93% (p=0.000 n=10)
ReturnTrieLayerVariable_VectorizedAlgorithm-32                                                843.9µ ±   1%     815.1µ ±   1%   -3.41% (p=0.000 n=10)
Uint64ListRootWithRegistryLimit/100k_balances-32                                              699.0µ ±   1%     689.4µ ±   1%   -1.37% (p=0.000 n=10)
StateMachine_trigger-32                                                                       78.47n ±   5%     77.34n ±   3%        ~ (p=0.190 n=10)
Value/100,000_shared_items-32                                                                 486.4µ ±   2%     474.0µ ±   2%   -2.53% (p=0.004 n=10)
Value/100,000_equal_individual_items-32                                                       1.131m ±   1%     1.133m ±   2%        ~ (p=0.579 n=10)
Value/100,000_different_individual_items-32                                                   3.248m ±   3%     3.334m ±   6%        ~ (p=0.089 n=10)
Value/100,000_shared_items_and_100,000_equal_appended_items-32                                567.9µ ±   1%     557.4µ ±   1%   -1.86% (p=0.000 n=10)
Value/100,000_shared_items_and_100,000_different_appended_items-32                            534.4µ ±   2%     540.9µ ±   1%        ~ (p=0.123 n=10)
Value/1,000,000_shared_items-32                                                               4.691m ±   6%     4.746m ±   2%        ~ (p=0.315 n=10)
Value/1,000,000_equal_individual_items-32                                                     9.142m ±   4%     8.646m ±   4%   -5.43% (p=0.007 n=10)
Value/1,000,000_different_individual_items-32                                                 56.41m ±   2%     55.62m ±   3%   -1.41% (p=0.002 n=10)
Value/1,000,000_shared_items_and_1,000,000_equal_appended_items-32                            4.699m ±   7%     4.578m ±   4%        ~ (p=0.739 n=10)
Value/1,000,000_shared_items_and_1,000,000_different_appended_items-32                        4.711m ±   7%     4.653m ±   5%        ~ (p=0.353 n=10)
Value/10,000,000_shared_items-32                                                              31.04m ±   4%     29.29m ±   7%   -5.63% (p=0.043 n=10)
Value/10,000,000_equal_individual_items-32                                                    85.43m ±   3%     84.65m ±   2%        ~ (p=0.052 n=10)
Value/10,000,000_different_individual_items-32                                                 3.036 ±   2%      2.959 ±   2%   -2.55% (p=0.007 n=10)
Value/10,000,000_shared_items_and_10,000,000_equal_appended_items-32                          33.29m ±  13%     32.72m ±   3%        ~ (p=0.353 n=10)
Value/10,000,000_shared_items_and_10,000,000_different_appended_items-32                      337.1m ±   2%     202.7m ±  56%  -39.88% (p=0.002 n=10)
Map_Concrete-32                                                                               29.61µ ±   1%     29.02µ ±   2%   -2.01% (p=0.008 n=10)
Map_Generic-32                                                                                31.59µ ±   1%     31.06µ ±   1%   -1.66% (p=0.000 n=10)
Map_GenericTx-32                                                                              18.79µ ±   1%     18.14µ ±   2%   -3.48% (p=0.000 n=10)
GenerateTrieFromItems-32                                                                      14.95µ ±   2%     14.46µ ±   3%   -3.27% (p=0.002 n=10)
InsertTrie_Optimized-32                                                                       8.714µ ±   4%     8.509µ ±   5%        ~ (p=0.052 n=10)
GenerateProof-32                                                                              1.734µ ±   2%     1.688µ ±   3%        ~ (p=0.052 n=10)
VerifyMerkleProofWithDepth-32                                                                 8.000µ ±   4%     7.751µ ±   3%   -3.12% (p=0.019 n=10)
Signature_Verify-32                                                                           1.137m ±   2%     1.132m ±  16%        ~ (p=0.315 n=10)
Signature_AggregateVerify-32                                                                  2.549m ±   0%     2.546m ±   1%        ~ (p=0.529 n=10)
SecretKey_Marshal-32                                                                          134.3n ±   6%     132.4n ±  10%        ~ (p=0.869 n=10)
PublicKeyFromBytes/cache_on-32                                                                189.2n ±   2%     185.1n ±   3%   -2.17% (p=0.037 n=10)
PublicKeyFromBytes/cache_off-32                                                               43.04µ ±   1%     42.98µ ±   1%        ~ (p=0.912 n=10)
HashKeccak256-32                                                                              842.1n ±   2%     839.6n ±   2%        ~ (p=0.393 n=10)
HashProto-32                                                                                  1.567µ ±   7%     1.533µ ±   8%        ~ (p=1.000 n=10)
ToBytes32-32                                                                                 0.2004n ±  10%    0.2009n ±  11%        ~ (p=0.897 n=10)
SafeCopyBytes/Copy_Bytes-32                                                                 0.02985n ±  12%   0.02143n ±  28%  -28.21% (p=0.029 n=10)
IntegerSquareRootBelow52Bits-32                                                               33.73n ±  13%     31.24n ±   7%   -7.40% (p=0.043 n=10)
IntegerSquareRootAbove52Bits-32                                                               64.98n ±   9%     63.22n ±   9%        ~ (p=0.631 n=10)
…